requirements.txt
# Core AI and LLM dependencies
litellm>=1.40.0
google-adk>=0.1.0

# Web framework dependencies (Streamlit replaces Gradio/FastAPI)
streamlit>=1.28.0
streamlit-chat>=0.1.1

# Environment and configuration
python-dotenv>=1.0.0

# Async and concurrency
asyncio-timeout>=4.0.0

# Type hints (included in Python 3.5+, but explicit for older versions)
typing-extensions>=4.0.0

# Data handling and export
pandas>=1.5.0
openpyxl>=3.0.0

# Standard library dependencies (usually included with Python)
# asyncio - built-in
# concurrent.futures - built-in
# json - built-in
# logging - built-in
# os - built-in
# queue - built-in
# threading - built-in

# Removed dependencies (no longer needed)
# gradio>=4.0.0  # Replaced by Streamlit
# fastapi>=0.100.0  # Not needed for Streamlit app
# uvicorn[standard]>=0.23.0  # Not needed for Streamlit app